+++ apps/lib/apps.c
@@ -25,6 +25,7 @@
 #endif
 #include <ctype.h>
 #include <errno.h>
+#include <openssl/macros.h>
 #include <openssl/err.h>
 #include <openssl/x509.h>
 #include <openssl/x509v3.h>