+++ fuzz/server.c
@@ -491,17 +491,21 @@ time_t time(time_t *t) TIME_IMPL(t)
 
 int FuzzerInitialize(int *argc, char ***argv)
 {
-    STACK_OF(SSL_COMP) *comp_methods;
-
     FuzzerSetRand();
     OPENSSL_init_crypto(OPENSSL_INIT_LOAD_CRYPTO_STRINGS | OPENSSL_INIT_ASYNC, NULL);
     OPENSSL_init_ssl(OPENSSL_INIT_LOAD_SSL_STRINGS, NULL);
     ERR_clear_error();
     CRYPTO_free_ex_index(0, -1);
     idx = SSL_get_ex_data_X509_STORE_CTX_idx();
-    comp_methods = SSL_COMP_get_compression_methods();
-    if (comp_methods != NULL)
-        sk_SSL_COMP_sort(comp_methods);
+
+#if !defined(OPENSSL_NO_COMP) && !defined(OPENSSL_NO_DEPRECATED_3_0)
+    {
+        STACK_OF(SSL_COMP) *comp_methods = SSL_COMP_get_compression_methods();
+
+        if (comp_methods != NULL)
+            sk_SSL_COMP_sort(comp_methods);
+    }
+#endif
 
     return 1;
 }