+++ ssl/ssl_lib.c
@@ -9,6 +9,7 @@
  * https://www.openssl.org/source/license.html
  */
 
+#include <internal/deprecated.h>
 #include <stdio.h>
 #include "ssl_local.h"
 #include "e_os.h"
@@ -3270,9 +3271,13 @@ SSL_CTX *SSL_CTX_new_ex(OSSL_LIB_CTX *libctx, const char *propq,
     if ((ret->ext.secure = OPENSSL_secure_zalloc(sizeof(*ret->ext.secure))) == NULL)
         goto err;
 
+#ifndef OPENSSL_NO_DEPRECATED_3_0
+# ifndef OPENSSL_NO_COMP
     /* No compression for DTLS */
     if (!(meth->ssl3_enc->enc_flags & SSL_ENC_FLAG_DTLS))
         ret->comp_methods = SSL_COMP_get_compression_methods();
+# endif
+#endif
 
     ret->max_send_fragment = SSL3_RT_MAX_PLAIN_LENGTH;
     ret->split_send_fragment = SSL3_RT_MAX_PLAIN_LENGTH;
@@ -4166,23 +4171,17 @@ const SSL_CIPHER *SSL_get_pending_cipher(const SSL *s)
     return s->s3.tmp.new_cipher;
 }
 
+#if !defined(OPENSSL_NO_COMP) && !defined(OPENSSL_NO_DEPRECATED_3_0)
 const COMP_METHOD *SSL_get_current_compression(const SSL *s)
 {
-#ifndef OPENSSL_NO_COMP
     return s->compress ? COMP_CTX_get_method(s->compress) : NULL;
-#else
-    return NULL;
-#endif
 }
 
 const COMP_METHOD *SSL_get_current_expansion(const SSL *s)
 {
-#ifndef OPENSSL_NO_COMP
     return s->expand ? COMP_CTX_get_method(s->expand) : NULL;
-#else
-    return NULL;
-#endif
 }
+#endif
 
 int ssl_init_wbio_buffer(SSL *s)
 {